+++ kernel/src/main/kotlin/eu/simonbinder/kart/kernel/utils/Flags.kt
@@ -1,25 +1,21 @@
 package eu.simonbinder.kart.kernel.utils
 
-import kotlin.experimental.and
-import kotlin.experimental.or
 import kotlin.reflect.KProperty
 
 interface HasFlags {
-    var flags: Byte
+    var flags: Int
 }
-
-// flag 0..7, where 0 is the least significant bit
 
 inline class PropertyFromFlag(private val index: Int) {
     operator fun getValue(thisRef: HasFlags, property: KProperty<*>): Boolean {
-        return ((thisRef.flags.toInt() shr index) and 1) == 1
+        return ((thisRef.flags shr index) and 1) == 1
     }
 
     operator fun setValue(thisRef: HasFlags, property: KProperty<*>, value: Boolean) {
         thisRef.flags = if (value) {
-            thisRef.flags or (1 shl index).toByte()
+            thisRef.flags or (1 shl index)
         } else {
-            thisRef.flags and (1 shl index).inv().toByte()
+            thisRef.flags and (1 shl index).inv()
         }
     }
 }
